changed last layer from Softmaxwithloss to Softmax

+++ yonatan/prepare_for_new_training.py
@@ -26,20 +26,22 @@ sets = {'train', 'cv', 'test'}
 for dress_type in sleeve_dress:
     print dress_type
     if os.path.isdir('/home/yonatan/resized_db_' + dress_type + '_dresses'):
-        if not os.listdir('/home/yonatan/resized_db_' + dress_type + '_dresses'):
-            print '\nempty'
-        #shutil.rmtree('/home/yonatan/db_' + dress_type + '_dresses')
-    else:
+        #if not os.listdir('/home/yonatan/resized_db_' + dress_type + '_dresses'):
+        #    print '\nempty'
+        shutil.rmtree('/home/yonatan/resized_db_' + dress_type + '_dresses')
+    '''else:
         #shutil.rmtree('/home/yonatan/db_' + dress_type + '_dresses')
         shutil.rmtree('/home/yonatan/resized_db_' + dress_type + '_dresses')
         #os.mkdir('/home/yonatan/db_' + dress_type + '_dresses')
         os.mkdir('/home/yonatan/resized_db_' + dress_type + '_dresses')
+    '''
 
 for set in sets:
     print set
-    shutil.rmtree('/home/yonatan/db_dresses_' + set + '_set')
-    #shutil.rmtree('/home/yonatan/dresses_' + set + '_set')
-    #os.mkdir('/home/yonatan/dresses_' + set + '_set')
-    os.mkdir('/home/yonatan/db_dresses_' + set + '_set')
-    #os.remove('/home/yonatan/db_dresses_' + set + '.txt')
+    if os.path.isdir('/home/yonatan/db_dresses_' + set + '_set'):
+        shutil.rmtree('/home/yonatan/db_dresses_' + set + '_set')
+        #shutil.rmtree('/home/yonatan/dresses_' + set + '_set')
+        #os.mkdir('/home/yonatan/dresses_' + set + '_set')
+        #os.mkdir('/home/yonatan/db_dresses_' + set + '_set')
+        #os.remove('/home/yonatan/db_dresses_' + set + '.txt')
